refactor: Move functions to functions.py
